refactor: import 경로 수정 및 엔티티 구조 개선
